requirements.txt
loguru
pydantic
fastmcp
awscli
boto3
pandas
python-dateutil
pytest